+++ app/src/main/java/woowacourse/shopping/data/network/ShoppingMockWebServer.kt
@@ -7,7 +7,7 @@ import okhttp3.mockwebserver.MockResponse
 import okhttp3.mockwebserver.MockWebServer
 import okhttp3.mockwebserver.RecordedRequest
 import woowacourse.shopping.constants.MockData
-import woowacourse.shopping.data.network.product.ProductDto
+import woowacourse.shopping.data.network.product.dto.ProductDto
 import woowacourse.shopping.domain.Product
 
 fun startMockWebServer(): MockWebServer {
