+++ app/src/main/java/woowacourse/shopping/data/network/cart/dto/Quantity.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.network.cart
+package woowacourse.shopping.data.network.cart.dto
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
